Switch to Devise.
As much as I do appreciate the original approach, I was finding we were coding up a lot of features that Devise provides for us - plus the additional wisdom of years of security patches and approaches.

I still need to rework the membership and profile sides of things, but this pulls out the bulk of the old approach, and puts Devise in its place. As a neat bonus, Devise’s password encryption is the same as `has_secure_password`, so everyone’s passwords will continue to work.