Default email template (used by Devise)
This was generated by Campaign Monitor, and so there’s a tonne of styling and markup there that I’m not quite sure about. But: they know what they’re doing when it comes to making emails look right in different email clients, so I’m going to leave it be.

I’m sure there are improvements to be made, of course, but this gives us a decent start, and is much closer in vibe to the current website design.