Create CM webhooks and listen to them.
This takes care of staying on top of sub/unsubs on the mailing lists to keep local flags up-to-date.

We’re not storing the webhook ids to refer to them later, but the odds of creating more than one are slim (as it’ll require CLI access to the app), and it’s not too hard to dive into a console and clean things up should mistakes get made.