Handle proxy registration.
This does not currently allow for noting votes for special resolutions tied to general meetings, but it’s otherwise solid enough.

The spec doesn’t check the actual signature - but I don’t think that’s a good enough reason to introduce the complexity of drawing in a canvas in a headless browser.