Simple registration field to stop spammers.
We’re getting hundreds of spam registrations each week - and while they’re not being confirmed (and thus don’t end up on our list of members), it’s still extra emails and database rows that I’d rather avoid.

So, this commit just adds another field to the registration form - and if it’s not filled in with ‘Ruby’ (any case, any part of the string - I like being flexible), then the registration is blocked.