Load CSS from an explicit file
While the CSS is being loaded via application.js in development, the CSS
tag isn't added to the page in production. While there would be a
solution to that problem, we don't care enough about loading the CSS
from the JS, and it's just easier to load the CSS from an explicit file.