Remove circular reference in Vite build options
By referencing the `application.js` entrypoint in
`build.rollupOptions.input.application`, a circular reference is created
in the asset manifest generated by Vite Ruby.

Asset helper tags, e.g. `vite_image_tag`, call
`ViteRuby::Manifest.path_for` to find the required asset. If the asset
doesn't exist, a `ViteRuby::MissingEntrypointError` is raised and the
error message includes a JSON representation of the manifest. Since the
manifest hash for the asset includes an "imports" key that references
the same entry in the manifest, a `JSON::NestingError` error is raised.

Annoyingly for debugging, Rollup only gets involved in production. The
assets will compile, and then an exception is raised when loading a
missing asset in production