Remove Vite build options
These options appear to either duplicate default or be unnecessary.

`build.minify` defaults to 'esbuild' for client build.
`build.cssMinify` defaults to the value of `build.minify`.
`build.outDir` defaults to `public/vite` for production builds.

We don't use jQuery in this project. It is used in `2019_survey.js`,
however, that file is not loaded, even on in `2019survey.html.erb`.

While `build.sourcemap` defaults to `true`, this is done to align with
(Rails default behaviour)[https://github.com/rails/webpacker/issues/769]