Fix circular reference in Vite/Rollup configuration (#320)
* Remove circular reference in Vite build options

By referencing the `application.js` entrypoint in
`build.rollupOptions.input.application`, a circular reference is created
in the asset manifest generated by Vite Ruby.

Asset helper tags, e.g. `vite_image_tag`, call
`ViteRuby::Manifest.path_for` to find the required asset. If the asset
doesn't exist, a `ViteRuby::MissingEntrypointError` is raised and the
error message includes a JSON representation of the manifest. Since the
manifest hash for the asset includes an "imports" key that references
the same entry in the manifest, a `JSON::NestingError` error is raised.

Annoyingly for debugging, Rollup only gets involved in production. The
assets will compile, and then an exception is raised when loading a
missing asset in production

* Remove Vite build options

These options appear to either duplicate default or be unnecessary.

`build.minify` defaults to 'esbuild' for client build.
`build.cssMinify` defaults to the value of `build.minify`.
`build.outDir` defaults to `public/vite` for production builds.

We don't use jQuery in this project. It is used in `2019_survey.js`,
however, that file is not loaded, even on in `2019survey.html.erb`.

While `build.sourcemap` defaults to `true`, this is done to align with
(Rails default behaviour)[https://github.com/rails/webpacker/issues/769]

* Remove explicit config for Stimulus HMR

This is enabled by default in ViteRails

* Remove server config for Vite

This does not appear to be a valid config. While Vite Ruby accepts
server configuration, they would be set in `vite.json`

* Exclude tmp from Rubocop analysis

`tmp` can be populated with server generated files and scratch pads. It
isn't a space used for team collaboration