Reduce memory required to perform an asset precompile
* Split application.js into multiple files

We are hitting our Heroku memory limit when compiling our assets. A lot
of memory is required to perform tasks like tree shaking, so we hope
that by splitting the application.js into multiple files, we can reduce
the maximum memory usage

* Removed unused icons.js file

* Limit maximum memory for node when building assets

We are hitting our Heroku memory limit when compiling our assets.
Limiting the maximum memory for node when performing a precompile should
force it to be more aggressive about garbage collection, just so long as
we ensure the memory limit is greater than the memory required to
compile any single file

In testing, I found that setting the memory limit to 384MB was enough to
compile without hitting the memory limit, but 256MB was not enough