Juli/fix safari video navigation bug (#335)
This fix was proposed by amp code. I'll test it in prod because I couldn't quite run the app locally in safari (which is something I need to fix too to be able to run these tests). It's a simple front end thing so it's safe to test it in prod